+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - [#3983](https://github.com/KronicDeth/intellij-elixir/pull/3983) [@sh41](https://github.com/sh41)
   - **Elixir's `Tools` menu actions are grouped under `Tools > Elixir`.**
+  - **New `Tools > Elixir > Sync Dependency Libraries` Action**. Rescans every content root's
+    `deps` and `_build`.
 
 - [#3954](https://github.com/KronicDeth/intellij-elixir/pull/3954) [@sh41](https://github.com/sh41)
   - **`~E` sigils now get EEx highlighting and completion.** `~H` injected HEEx and `~L` injected EEx,